Moved AppEngineFacet#installAppEngineFacet and AppEngineFacet#installAppEngineRuntime into FacetInstallDelegate
