+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject.maven/src/com/google/cloud/tools/eclipse/appengine/newproject/maven/CreateMavenBasedAppEngineStandardProject.java
@@ -19,6 +19,7 @@ import org.eclipse.wst.common.project.facet.core.IFacetedProject;
 import org.eclipse.wst.common.project.facet.core.ProjectFacetsManager;
 
 import com.google.cloud.tools.eclipse.appengine.facets.AppEngineFacet;
+import com.google.cloud.tools.eclipse.appengine.facets.FacetInstallDelegate;
 
 import java.lang.reflect.InvocationTargetException;
 import java.text.MessageFormat;
@@ -76,8 +77,8 @@ public class CreateMavenBasedAppEngineStandardProject extends WorkspaceModifyOpe
     for (IProject project : archetypeProjects) {
       IFacetedProject facetedProject = ProjectFacetsManager.create(
           project, true, loopMonitor.newChild(1));
-      AppEngineFacet.installAppEngineFacet(facetedProject, loopMonitor.newChild(1));  
-      AppEngineFacet.installAppEngineRuntime(facetedProject, loopMonitor.newChild(1));  
+      FacetInstallDelegate.installAppEngineFacet(facetedProject, loopMonitor.newChild(1));
+      FacetInstallDelegate.installAppEngineRuntime(facetedProject, loopMonitor.newChild(1));
     }
     
     /*
